requirements.txt
streamlit
langchain
langchain-mistralai
python-dotenv
requests